change return type of centAmount from int to long. (#63)
* change return type of centAmount from int to long.

* add CHANGELOG.MD

* update changelog; add breaking changes section

+++ commercetools.NET/Common/Money.cs
@@ -14,7 +14,7 @@ namespace commercetools.Common
         public string CurrencyCode { get; set; }
 
         [JsonProperty(PropertyName = "centAmount")]
-        public int? CentAmount { get; set; }
+        public long? CentAmount { get; set; }
 
         #endregion
